Set up the same subid mappings for all images
Use the same subid mappings for the image's custom user for each of the
images: ranges 1:999 and 1001:64535.  This specifically changes
skopeo's "skopeo" user, which used to have 100000:65536 assigned.

Add a 1:65535 range for "root", for cases where the image is run without
CAP_SYS_ADMIN, but it _is_ run as UID 0 with CAP_SETFCAP, and can use
unshare(2) to set up a user namespace.

Fixes #46

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>