Make SOURCE_DATE_EPOCH available for builds
If we're passed a SOURCE_DATE_EPOCH build arg, declare it so that it
gets passed to rpm, and use it to rewrite the dates in the dnf
transaction database before flushing its write-ahead journal by deleting
and restarting it.

Remove hawkey logs, "countme" files, and the generated machine-id and
rng seed.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>